myproject/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
